+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 
 - Viewport Height component no longer sets a `height` except for IE  
 - Improve Tooltip performance
+- Improve Parallax performance
 
 ### Removed